requirements.txt
opencv-python>=4.5.0
albumentations>=1.3.0
numpy>=1.19.0
scikit-image>=0.19.0
pyyaml>=5.1.0 